% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{check_for_invalid_param_names}
\alias{check_for_invalid_param_names}
\title{check_for_invalid_param_names}
\usage{
check_for_invalid_param_names(ls.parameters)
}
\arguments{
\item{ls.parameters}{a list of easypop parameter names and values}
}
\description{
in a list of param names values,
we compare each name in the list to
the list of valid easypop parameter names
with a call to stop and a message if any names
are not valid
}
