% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{convert.param.val.from.string}
\alias{convert.param.val.from.string}
\title{convert.param.val.from.string
given a parameter name, convert
the given value to one of numeric, integer,
or character, using the master parameter list}
\usage{
convert.param.val.from.string(s.param.name, s.value)
}
\arguments{
\item{s.param.name}{character easypop config file parameter name}

\item{s.value}{character (string) value to be converted

Assumes the s.param.name}
}
\description{
convert.param.val.from.string
given a parameter name, convert
the given value to one of numeric, integer,
or character, using the master parameter list
}
