\name{easypopr-package}
\alias{easypopr-package}
\alias{easypopr}
\docType{package}
\title{
\packageTitle{easypopr}
}
\description{
\packageDescription{easypopr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{easypopr}
\packageIndices{easypopr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{easypopr}

Maintainer: \packageMaintainer{easypopr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
