% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.utils.R
\name{format_snpR}
\alias{format_snpR}
\title{Format easypopr genepop outfiles into snpRdata objects}
\usage{
format_snpR(files, verbose = FALSE)
}
\arguments{
\item{files}{Either a vector of file names pointing to easypop/genepop
output files ending in either '.gen' or '.genepop' or a run info table
like that produced by \code{\link{configure_multiple_easypop_runs}}.}

\item{verbose}{logical, default FALSE. If TRUE, prints progress statements.}
}
\value{
A nested list containing snpRdata objects in '$dat' and possibly
'$meta' containing run metadata if provided an run info table.
}
\description{
Generates data files for the 'snpR' package given a vector of file names or
config outputs from \code{\link{configure_multiple_easypop_runs}}. Requires
'snpR' to be installed.
}
\examples{
\dontshow{if ("snpR" \%in\% rownames(utils::installed.packages())) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# pull in example config file
s.config.file = system.file( "extdata", 
                             "demo.standard.cfg", 
                             package = "easypopr") 

# run for two different numbers of populations
l.params.to.vary = list(
  "number_populations" = c(5, 10))

# run
info <- 
  configure_multiple_easypop_runs( l.settings =  l.params.to.vary, 
                                   s.starting.config.file = s.config.file, 
                                   s.filebase = "test_pops",
                                   b.run = TRUE )

# pull into snpR format
d <- format_snpR(info)

# check results
d$meta # metadata for each object
snpR::summarize_facets(d$dat[[1]], "genepop_pop") # five pops
snpR::summarize_facets(d$dat[[3]], "genepop_pop") # ten pops

# can also be run with a vector of file names:
d <- format_snpR(unlist(info$outfiles))
\dontshow{\}) # examplesIf}
}
