% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{get.genetic.parameters}
\alias{get.genetic.parameters}
\title{get.genetic.parameters}
\usage{
get.genetic.parameters()
}
\description{
prompts user for genetic parameters and returns user-entered values as a list.
}
