% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{get.migration.parameters}
\alias{get.migration.parameters}
\title{get.migration.parameters}
\usage{
get.migration.parameters(g2sex, gnbpop)
}
\arguments{
\item{g2sex, }{boolean, 1 or 0, whether there are 2 sexes.}

\item{gnbpop, }{integer, number of populations.}
}
\description{
prompts user for migration scheme parameters and returns user-entered values as a list.
}
