% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{get.migration.scheme.details}
\alias{get.migration.scheme.details}
\title{get.migration.scheme.details}
\usage{
get.migration.scheme.details(g2sex, gnbpop, b.second.scheme)
}
\arguments{
\item{g2sex, }{boolean, 1 or 0, whether there are 2 sexes.}

\item{gnbpop, }{integer, number of populations.}
}
\description{
prompts user for details of migration scheme parameters and returns user-entered values as a list.
this function is called only if the number of pops exceeds 2 (else they are trivially assigned island in 
the fx, get.migration.parameters.
}
