% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{get.population.parameters}
\alias{get.population.parameters}
\title{get.population.parameters}
\usage{
get.population.parameters(gploidy, g2sex)
}
\arguments{
\item{gploidy, }{integer giving ploidy}

\item{g2sex, }{boolean, 1 or 0, whether there are 2 sexes.}
}
\description{
prompts user for population parameters and returns user-entered values as a list.
}
