% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{get.reproduction.parameters}
\alias{get.reproduction.parameters}
\title{get.reproduction.parameters}
\usage{
get.reproduction.parameters()
}
\description{
prompts user for reproduction parameters and returns user-entered values as a list.
}
