% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{get.selfing.parameters}
\alias{get.selfing.parameters}
\title{get.selfing.parameters}
\usage{
get.selfing.parameters()
}
\description{
prompt user for selfing parameters, return a list with user-entered values.
}
