% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{locate_easypop}
\alias{locate_easypop}
\title{Manually locate an EASYPOP executable}
\usage{
locate_easypop(path)
}
\arguments{
\item{path}{character, a path to the EASYPOP executable.}
}
\description{
Manually locate an EASYPOP executable
}
