% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{make.list.from.cfg.lines}
\alias{make.list.from.cfg.lines}
\title{make.list.from.cfg.lines}
\usage{
make.list.from.cfg.lines(v.lines)
}
\arguments{
\item{v.lines}{vector of character strings,
each a config file line}
}
\description{
from a vector of easypop config
file lines, make a list whose
names are param names, and whose
values are the corresponding param vals
}
