% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.analysis.R
\name{plot.equ.values}
\alias{plot.equ.values}
\title{plot.equ.values}
\usage{
\method{plot}{equ.values}(s.colname, ldf.data.frames)
}
\arguments{
\item{s.colname}{a column name from the equ file's first line}

\item{ldf.data.frames}{a list of data frames, as created by calling read.table
on an *.equ file.}
}
\description{
for each data frame given in arg ldf.data.frames, list of items, each representing 
an ep output *equ file,  a line is plotted showing the per-generation values for the
quantity (e.g. Fst), associated with the column name given by arg s.colname
and returns them in a vector length i.num.values, of user-entered values
}
