% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.analysis.R
\name{plot_easypop_replicate_equ_means}
\alias{plot_easypop_replicate_equ_means}
\title{plot_easypop_replicate_equ_means}
\usage{
plot_easypop_replicate_equ_means(vs.config.files, s.colname)
}
\arguments{
\item{vs.config.files}{vector of names of ep config files}

\item{s.colname}{names one of the columns of an equ file,
as given in the first line of an equ file, one of,
Ho, Hs, Ht, Fis, Fst, or Fit.}
}
\description{
for each easypop config file given in the vector arg,
Plot the per-generation, replicate mean values for the 
equ output files, the column given by the s.colname arg.
The name_of_file entry  in each config file is used to
locate the *equ files
}
