% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.analysis.R
\name{plot_ep_replicate_equ_means}
\alias{plot_ep_replicate_equ_means}
\title{plot_ep_replicate_equ_means}
\usage{
plot_ep_replicate_equ_means(vs.config.files, s.colname)
}
\arguments{
\item{vs.config.files}{vector of names of ep config files}

\item{s.colname}{names one of the columns of an equ file,
as given in the first line of an equ file.}
}
\description{
for each easypop config file given in the vector arg,
plot the per-generation, replicate mean values for the 
equ output files, the column given by the s.colname arg 
the name_of_file entry  in each config file is used to
locate the *equ files
}
