% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{print.param.list}
\alias{print.param.list}
\title{print.param.list
prints the param name-value pairs to the file named in the args.}
\usage{
\method{print}{param.list}(lv.param.list, s.filename)
}
\arguments{
\item{lv.param.list}{list whose names are ep param names and values are the corresponding values}

\item{s.filename}{name of the file to which to write the name-value pairs}
}
\description{
print.param.list
prints the param name-value pairs to the file named in the args.
}
