% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{prompt.for.values.and.return.user.entries}
\alias{prompt.for.values.and.return.user.entries}
\title{prompt.for.values.and.return.user.entries}
\usage{
\method{prompt}{for.values.and.return.user.entries}(s.prompt, i.num.values, s.type.of.values, v.valid.range)
}
\arguments{
\item{s.prompt, }{printed to console on each iteration to get another value}

\item{i.num.values, }{number of entries to get from the user.}

\item{s.type.of.values, }{string, one of "integer", "numeric", or "character"}

\item{v.valid.range}{min, max values (inclusive) of valid values. If type "character", then NULL or list of valid strings)}
}
\description{
prompts user at the console for parameter values, stores
and returns them in a vector length i.num.values, of user-entered values
}
