% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{read.config.file}
\alias{read.config.file}
\title{read.config.file
Reads in an ep config file as a list}
\usage{
read.config.file(s.file)
}
\arguments{
\item{s.file}{names an ep config file}
}
\description{
read.config.file
Reads in an ep config file as a list
}
