% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{run.easypop}
\alias{run.easypop}
\title{run.easypop}
\usage{
run.easypop(s.file.name)
}
\arguments{
\item{s.file.name, }{string giving the name of an existing file holding EASYPOP simulation parameters}
}
\description{
Runs an EASYPOP simulation using the parameters given in the file
named by the argument.
}
