% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{setup.easypop}
\alias{setup.easypop}
\title{setup.easypop}
\usage{
setup.easypop(s.file.name)
}
\arguments{
\item{s.file.name, }{string giving the name of a new file to which the program can
write the parameter values entered at the prompts.}
}
\description{
Prompts the user for parameter values used to run an EASYPOP simulation. 
and writes the values to a file named by the argument.
}
