% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{try.to.convert}
\alias{try.to.convert}
\title{try.to.convert
given a character-type, convert it to the type
named by the given target type}
\usage{
try.to.convert(s.string.to.convert, s.target.type)
}
\arguments{
\item{s.string.to.convert}{character string to be converted}

\item{s.target.type}{string naming a type, one of "numeric",
"integer", or (trivially) "character"
we make this fx more extensible by returning NULL
if the converstion fails, letting the caller handle failures.}
}
\description{
try.to.convert
given a character-type, convert it to the type
named by the given target type
}
