% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{write.config.file}
\alias{write.config.file}
\title{write.config.file
Writes an ep config file whose
parameter names are those of the arg list,
and values are the list's values.  These are
written to the file named by the 2nd arg.}
\usage{
write.config.file(ls.params.and.values, s.file)
}
\arguments{
\item{ls.params.and.values}{a list whose names are ep param names
and whose values are their corrsponding values.}

\item{s.file}{names the file to be written
The file name should not refer to an existing file}
}
\value{
a list whose names are easypop parameter names and whose
values are their associated values.
}
\description{
write.config.file
Writes an ep config file whose
parameter names are those of the arg list,
and values are the list's values.  These are
written to the file named by the 2nd arg.
}
