% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.utils.R
\name{configure_multiple_easypop_runs}
\alias{configure_multiple_easypop_runs}
\title{configure_multiple_easypop_runs}
\usage{
configure_multiple_easypop_runs(
  l.settings,
  s.starting.config.file,
  s.filebase,
  b.run = FALSE
)
}
\arguments{
\item{l.settings}{a list whose names are easypop config file parameter names,
and whose values are vectors of settings (values) for each parameter}

\item{s.starting.config.file}{names an easypop config file to be used as the basis
for the simulations, with revised settings as given by the first arg}

\item{b.run}{optional, default = FALSE, if set to TRUE, as each config is created,
a simulation is run based on the config}
}
\description{
given a list of parameter names, each with a vector of values,
for each combination of settings, create an easypop configuration file
and, optionally run each file as it is created.  After all configurations
are written (and, optionally, run), the program writes  a tabular file with
the given filebase and with extension, ".config.key.tsv"  This table gives
in it's first column a config file number, and in columns 2..N, the values
for the params that were in the l.settings list (see parameter descriptions).
}
