% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.analysis.R
\name{plot_easypop_replicate_equ_values}
\alias{plot_easypop_replicate_equ_values}
\title{plot_easypop_replicate_equ_values}
\usage{
plot_easypop_replicate_equ_values(s.colname, ldf.data.frames)
}
\arguments{
\item{s.colname}{a column name from the equ file's first line}

\item{ldf.data.frames}{either a list of data frames, as created by calling
\code{\link[utils]{read.table}} on a set of *.equ file, or a vector of file
paths to each .equ file you wish to read. The latter can be generated with
the \code{\link[base]{file.path}} using the \code{full.names = TRUE}
argument.}
}
\description{
for each data frame given in arg ldf.data.frames, list of items, each
representing an ep output *equ file,  a line is plotted showing the
per-generation values for the quantity (e.g. Fst), associated with the column
name given by arg s.colname and returns them in a vector length i.num.values,
of user-entered values
}
