% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.params.R
\name{read_parameters_from_file}
\alias{read_parameters_from_file}
\title{read_parameters_from_file
From an existing easypop config file,}
\usage{
read_parameters_from_file(s.file)
}
\arguments{
\item{s.file}{names an existing easypop configuration file}
}
\value{
a list whose names are the parameter names, and values
are the associated parameter values
}
\description{
get a list whose names are the parameter names
and whose values are the associated parameter values.
}
